+++ tests/test_ipv6.py
@@ -1,6 +1,6 @@
 import stomp
 from stomp.listener import TestListener
-from stomp import logging
+from stomp import log
 from .testutils import *
 
 
@@ -20,7 +20,7 @@ def conn():
 class TestIP6(object):
     def test_ipv6_send(self, conn):
         if not is_inside_travis():
-            logging.info("running ipv6 test")
+            log.info("running ipv6 test")
             timestamp = time.strftime("%Y%m%d%H%M%S")
             queuename = "/queue/testipv6-%s" % timestamp
             conn.subscribe(destination=queuename, id=1, ack="auto")