+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 * Update to docopt-ng (as per issue https://github.com/jasonrbriggs/stomp.py/issues/442)
 * Pass ssl version from ssl_params (https://github.com/jasonrbriggs/stomp.py/pull/440)
 * Use single thread for IO when using TLS (https://codeberg.org/jasonrbriggs/stomp.py/pulls/452, original github issue: https://github.com/jasonrbriggs/stomp.py/issues/450)
-*  Added SSL error message from cert validation to log warning (https://codeberg.org/jasonrbriggs/stomp.py/pulls/449)
-
+* Added SSL error message from cert validation to log warning (https://codeberg.org/jasonrbriggs/stomp.py/pulls/449)
+* Rename stomp.logging to stomp.log, since python doesn't "like" overloading the logging module
 
 
 ## Version 8.2.0 - Oct 2024